@@ -16,41 +16,38 @@
     "upload": "yarn tsx src/maintenance/upload.ts"
   },
   "dependencies": {
-    "itty-router": "^2.6.1",
-    "itty-router-extras": "^0.4.2",
+    "itty-router": "^5.0.22",
     "json-logic-js": "^2.0.2",
     "json5": "^2.2.3",
     "path-browserify": "^1.0.1",
-    "semver": "^7.5.2",
+    "semver": "^7.7.2",
     "zod": "^3.19.1"
   },
   "devDependencies": {
     "@actions/core": "^1.9.1",
     "@actions/github": "^5.0.3",
-    "@cloudflare/workers-types": "^4.20240821.1",
-    "@tsconfig/node16": "^1.0.3",
-    "@types/eslint": "^8.4.6",
-    "@types/itty-router-extras": "^0.4.0",
+    "@cloudflare/workers-types": "^4.20250910.0",
+    "@types/eslint": "^9.6.1",
     "@types/json-logic-js": "^2.0.1",
     "@types/node": "^22.18.1",
     "@types/path-browserify": "^1.0.0",
-    "@types/prettier": "^2.7.0",
-    "@types/semver": "^7.3.12",
-    "@types/source-map-support": "^0.5.6",
-    "@typescript-eslint/eslint-plugin": "^5.37.0",
-    "@typescript-eslint/parser": "^5.37.0",
+    "@types/semver": "^7.7.1",
+    "@types/source-map-support": "^0.5.10",
+    "@typescript-eslint/eslint-plugin": "^8.43.0",
+    "@typescript-eslint/parser": "^8.43.0",
     "@zwave-js/firmware-integrity": "^1.0.2",
     "ava": "^4.3.3",
-    "axios": "^0.27.2",
     "esbuild": "^0.15.7",
     "esbuild-register": "^3.3.3",
-    "eslint": "^8.23.1",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "prettier": "^2.7.1",
-    "prettier-plugin-organize-imports": "^3.1.1",
+    "eslint": "^9.35.0",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-prettier": "^5.5.4",
+    "ky": "^1.10.0",
+    "prettier": "^3.6.2",
+    "prettier-plugin-organize-imports": "^4.2.0",
     "tsx": "^3.13.0",
     "typescript": "~5.8.3",
+    "typescript-eslint": "^8.43.0",
     "wrangler": "^3.73.0",
     "zod-to-json-schema": "^3.17.1"
   },
