@@ -1,4 +1,4 @@
-import { error } from "itty-router-extras";
+import { error } from "itty-router";
 
 /** Constant-time string comparison */
 export function safeCompare(expected: string, actual: string): boolean {
@@ -19,21 +19,21 @@ export function safeCompare(expected: string, actual: string): boolean {
 
 export function clientError(
 	message: BodyInit | Record<string, any> | undefined,
-	code: number = 400
+	code: number = 400,
 ): Response {
 	return error(code, message);
 }
 
 export function serverError(
 	message: BodyInit | Record<string, any>,
-	code: number = 500
+	code: number = 500,
 ): Response {
 	return error(code, message);
 }
 
 export type RequestWithProps<
 	U extends Record<string, unknown>[],
-	T extends Request = Request
+	T extends Request = Request,
 > = U extends [infer First, ...infer Rest extends Record<string, unknown>[]]
 	? RequestWithProps<Rest, T & First>
 	: T;
