@@ -1,5 +1,5 @@
-import zodToJsonSchema from "zod-to-json-schema";
-import { configSchema } from "../lib/configSchema";
+import { zodToJsonSchema } from "zod-to-json-schema";
+import { configSchema } from "../lib/configSchema.js";
 
 const jsonSchema = zodToJsonSchema(configSchema, "firmwareConfig");
 console.log(JSON.stringify(jsonSchema, null, "\t"));
